pyproject.toml
[project]
name = "upnm-fl-cvd"
version = "0.1.0"
description = "Federated Learning System for Cardiovascular Disease Prediction"
authors = [{name = "UPNM Team"}]
dependencies = [
    "scikit-learn>=1.3.0",
    "flwr>=1.7.0",
    "pandas>=2.0.3",
    "numpy>=1.24.3",
    "joblib>=1.3.2",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
]